whisper_api/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
torch
